wip: create class component
