+++ app/javascript/components/custom_form.jsx
@@ -1,27 +1,12 @@
 import React, { Component } from "react";
-import Form from './form'
-import { render } from "react-dom";
-
-const schema = {
-  title: "Todo",
-  type: "object",
-  required: ["title"],
-  properties: {
-    title: {type: "string", title: "Title", default: "A new task"},
-    done: {type: "boolean", title: "Done?", default: false}
-  }
-};
-
-const log = (type) => console.log.bind(console, type);
+import ClassForm from './class_form'
 
 export default class CustomForm extends Component {
 
   render() {
     return (
-      <Form schema={schema}
-      onChange={log("changed")}
-      onSubmit={log("submitted")}
-      onError={log("errors")} />
+      <ClassForm
+      />
     );
   }
 }
