+++ app/views/hyrax/my/m3_profiles/new.html.erb
@@ -10,9 +10,10 @@
           <a href="#description" role="tab" data-toggle="tab"><%= t('.tabs.description') %></a>
         </li>
       </ul>
-      <h1>REACT</h1>
-      <%= react_component("HelloWorld", { greeting: "Hello from react-rails." }) %>
-      <%= react_component("custom_form") %>
+      <div class="panel-body">
+        <h1>REACT</h1>
+        <%= react_component("custom_form") %>
+      </div>
 
       <%#= simple_form_for @m3_profile, url: hyrax.my_m3_profiles_path, method: :post do |form| %>
         <%#= render 'form', m3_profile: @m3_profile, form: form %>
